scripts/requirements-local.txt
# Minimal deps for local YouTube ingestion
youtube-transcript-api==1.2.4
yt-dlp>=2025.1.15
openai>=1.59.9
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pgvector>=0.3.6
llama-index-core>=0.12.12
python-dotenv>=1.0.1
pydantic-settings>=2.7.1
